requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-dotenv
pymysql
cryptography
faker
openai
pytest
httpx
pytest-asyncio
slowapi
